==> 06_graphics_tables.Rmd <==
---
title: "graphics_tables"
author: "Leandro Pereira Garcia"
date: "October 10, 2019"
output: 
  word_document: 
    reference_docx: TESEv.06.docx
    fig_caption: yes
    fig_height: 8
    fig_width: 12
    toc: no
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message=FALSE, warning=FALSE)
options(scipen=999)
set.seed(233)
```


```{r include=FALSE}
#Libraries
library(readr)
library(tidyverse)
library(ggplot2)
library(sf)
library(spData)
library(ggpubr)
library(RColorBrewer)
library(corrplot)
library(ggcorrplot)
library(reshape2)
```


```{r include=FALSE}
#Databases
base_efi <- read_csv("bases/efficiency_analysis.csv")
```



```{r include=FALSE}
#Importing the world map
WorldData <- map_data('world')
WorldData %>% filter(region != "Antarctica") -> WorldData
WorldData <- fortify(WorldData)


base_efi$region_map <- base_efi$LOCATION
base_efi[which(base_efi$LOCATION == "Congo, Rep."), 60] <- "Republic of Congo"
base_efi[which(base_efi$LOCATION == "Egypt, Arab Rep."), 60] <- "Egypt"
base_efi[which(base_efi$LOCATION == "Korea, Rep."), 60] <- "South Korea"
base_efi[which(base_efi$LOCATION == "North Macedonia"), 60] <- "Macedonia"
base_efi[which(base_efi$LOCATION == "Russian Federation"), 60] <- "Russia"
base_efi[which(base_efi$LOCATION == "Trinidad and Tobago"), 60] <- "Trinidad"
base_efi[which(base_efi$LOCATION == "United Kingdom"), 60] <- "UK"
base_efi[which(base_efi$LOCATION == "United States"), 60] <- "USA"
base_efi[which(base_efi$LOCATION == "Brunei Darussalam"), 60] <- "Brunei"
base_efi[which(base_efi$LOCATION == "Congo, Dem. Rep."), 60] <- "Democratic Republic of the Congo"
base_efi[which(base_efi$LOCATION == "Gambia, The"), 60] <- "Gambia"
base_efi[which(base_efi$LOCATION == "Cote d'Ivoire"), 60] <- "Ivory Coast"
base_efi[which(base_efi$LOCATION == "Iran, Islamic Rep."), 60] <- "Iran"
base_efi[which(base_efi$LOCATION == "Kyrgyz Republic"), 60] <- "Kyrgyzstan"
base_efi[which(base_efi$LOCATION == "Lao PDR"), 60] <- "Laos"
base_efi[which(base_efi$LOCATION == "Slovak Republic"), 60] <- "Slovakia"
base_efi[which(base_efi$LOCATION == "Yemen, Rep."), 60] <- "Yemen"


base_map <- base_efi
base_map <- merge(base_map, WorldData, by.x = "region_map", by.y = "region", all.x = T)

```

```{r fig.width=16, fig.height=10}
#Neontal, 28d<U5, Public Expenditure, Proportion of exp distribution
myPalette_rev <- colorRampPalette(brewer.pal(11, "Spectral")) #Increasing
myPalette <- colorRampPalette(rev(brewer.pal(11, "Spectral")))  #Decreasing

base_mapa_neo <- dplyr::select(base_map, c(region_map, MEAN_RATE_NEO))
plot1 <- ggplot()+
        geom_map(data=WorldData, map=WorldData,
                  aes(x=long, y=lat, group=group, map_id=region),
                  fill="white", colour="#7f7f7f", size=0.5)+
        geom_map(data=base_mapa_neo, map=WorldData,
                  aes(fill= MEAN_RATE_NEO, map_id=region_map),
                  colour="#7f7f7f", size=0.5)+
        coord_map("rectangular", lat0=0, xlim=c(-190,190), ylim=c(-60, 90))+
        labs(fill="Mean neo 2016-2017", x="", y="")+
        theme_bw()+
        theme(panel.border = element_blank())+
        scale_fill_gradientn(colours = myPalette(20), 
                             labels = round(quantile(base_mapa_neo$MEAN_RATE_NEO, c(.05,.95)),3), 
                             breaks = round(quantile(base_mapa_neo$MEAN_RATE_NEO, c(.05,.95)),3))

base_mapa_neo_u5 <- dplyr::select(base_map, c(region_map, MEAN_RATE_NEO_U5))
plot2 <- ggplot()+
        geom_map(data=WorldData, map=WorldData,
                  aes(x=long, y=lat, group=group, map_id=region),
                  fill="white", colour="#7f7f7f", size=0.5)+
        geom_map(data=base_mapa_neo_u5, map=WorldData,
                  aes(fill= MEAN_RATE_NEO_U5, map_id=region_map),
                  colour="#7f7f7f", size=0.5)+
        coord_map("rectangular", lat0=0, xlim=c(-190,190), ylim=c(-60, 90))+
        labs(fill="Mean 28d<5y 2016-2017", x="", y="")+
        theme_bw()+
        theme(panel.border = element_blank())+
        scale_fill_gradientn(colours = myPalette(20), 
                             labels = round(quantile(base_mapa_neo_u5$MEAN_RATE_NEO_U5, c(.05,.95)),3), 
                             breaks = round(quantile(base_mapa_neo_u5$MEAN_RATE_NEO_U5, c(.05,.95)),3))

base_mapa_public <- dplyr::select(base_map, c(region_map, PUBLIC_EXP_LAGGED2))
plot3 <- ggplot()+
        geom_map(data=WorldData, map=WorldData,
                  aes(x=long, y=lat, group=group, map_id=region),
                  fill="white", colour="#7f7f7f", size=0.5)+
        geom_map(data=base_mapa_public, map=WorldData,
                  aes(fill= PUBLIC_EXP_LAGGED2, map_id=region_map),
                  colour="#7f7f7f", size=0.5)+
        coord_map("rectangular", lat0=0, xlim=c(-190,190), ylim=c(-60, 90))+
        labs(fill="Mean pub exp 2013-2014", x="", y="")+
        theme_bw()+
        theme(panel.border = element_blank())+
        scale_fill_gradientn(colours = myPalette_rev(20), 
                             labels = round(quantile(base_mapa_public$PUBLIC_EXP_LAGGED2, c(.05,.95)),0), 
                             breaks = round(quantile(base_mapa_public$PUBLIC_EXP_LAGGED2, c(.05,.95)),0))

base_mapa_prop <- dplyr::select(base_map, c(region_map, PROP_PUBLIC_HEALTH_EXP_LAGGED2))
plot4 <- ggplot()+
        geom_map(data=WorldData, map=WorldData,
                  aes(x=long, y=lat, group=group, map_id=region),
                  fill="white", colour="#7f7f7f", size=0.5)+
        geom_map(data=base_mapa_prop, map=WorldData,
                  aes(fill= PROP_PUBLIC_HEALTH_EXP_LAGGED2, map_id=region_map),
                  colour="#7f7f7f", size=0.5)+
        coord_map("rectangular", lat0=0, xlim=c(-190,190), ylim=c(-60, 90))+
        labs(fill="Mean prop exp health 2013-2014", x="", y="")+
        theme_bw()+
        theme(panel.border = element_blank())+
        scale_fill_gradientn(colours = myPalette_rev(20), 
                             labels = round(quantile(base_mapa_prop$PROP_PUBLIC_HEALTH_EXP_LAGGED2, c(.05,.95)),2), 
                             breaks = round(quantile(base_mapa_prop$PROP_PUBLIC_HEALTH_EXP_LAGGED2, c(.05,.95)),2))

ggarrange(plot1, plot2, plot3,plot4, labels = c("A", "B", "C", "D"),
           common.legend = F,legend = "bottom")

```

```{r fig.width=12, fig.height=16}
base_efi[which(base_efi$INCOME_CLASS == "L"), 2] <- "Low"
base_efi[which(base_efi$INCOME_CLASS == "LM"), 2] <- "Low-Median"
base_efi[which(base_efi$INCOME_CLASS == "UM"), 2] <- "Upper-Median"
base_efi[which(base_efi$INCOME_CLASS == "H"), 2] <- "High"
base_efi$INCOME_CLASS <- factor(base_efi$INCOME_CLASS, levels = c("Low", "Low-Median", "Upper-Median", "High"))
base_efi_poor <- subset(base_efi, base_efi$INCOME_CLASS %in% c("Low", "Low-Median"))
```

O país pobre com a menor média na taxa de mortalidade neonatal  foi `r base_efi_poor[which(base_efi_poor$MEAN_RATE_NEO == min(base_efi_poor$MEAN_RATE_NEO)),1]`: `r round(base_efi_poor[which(base_efi_poor$MEAN_RATE_NEO == min(base_efi_poor$MEAN_RATE_NEO)),3],3)`. O com maior foi  `r base_efi_poor[which(base_efi_poor$MEAN_RATE_NEO == max(base_efi_poor$MEAN_RATE_NEO)),1]`, com média de `r round(base_efi_poor[which(base_efi_poor$MEAN_RATE_NEO == max(base_efi_poor$MEAN_RATE_NEO)),3],3)`. Com relação à taxa de mortalidade de crianças com idade entre 28 dias e cinco anos, o pais com a menor foi `r base_efi_poor[which(base_efi_poor$MEAN_RATE_NEO_U5 == min(base_efi_poor$MEAN_RATE_NEO_U5)),1]` e com a maior `r base_efi_poor[which(base_efi_poor$MEAN_RATE_NEO_U5 == max(base_efi_poor$MEAN_RATE_NEO_U5)),1]`, com médias de `r round(base_efi_poor[which(base_efi_poor$MEAN_RATE_NEO_U5 == min(base_efi_poor$MEAN_RATE_NEO_U5)),4],3)` e `r round(base_efi_poor[which(base_efi_poor$MEAN_RATE_NEO_U5 == max(base_efi_poor$MEAN_RATE_NEO_U5)),4],3)`, respectivamente.
Com relação à média dos gastos públicos (tratamento), O país pobre com a maior média de gastos foi `r base_efi_poor[which(base_efi_poor$PUBLIC_EXP_LAGGED2 == max(base_efi_poor$PUBLIC_EXP_LAGGED2)),1]`, US\$`r round(base_efi_poor[which(base_efi_poor$PUBLIC_EXP_LAGGED2 == max(base_efi_poor$PUBLIC_EXP_LAGGED2)),9],2)`; o com amenor foi `r base_efi_poor[which(base_efi_poor$PUBLIC_EXP_LAGGED2 == min(base_efi_poor$PUBLIC_EXP_LAGGED2)),1]`: US\$`r round(base_efi_poor[which(base_efi_poor$PUBLIC_EXP_LAGGED2 == min(base_efi_poor$PUBLIC_EXP_LAGGED2)),9],2)`. A proporção dos gastos públicos destinados à saúde foi maior em  `r base_efi_poor[which(base_efi_poor$PROP_PUBLIC_HEALTH_EXP_LAGGED2 == max(base_efi_poor$PROP_PUBLIC_HEALTH_EXP_LAGGED2)),1]`, `r round(base_efi_poor[which(base_efi_poor$PROP_PUBLIC_HEALTH_EXP_LAGGED2 == max(base_efi_poor$PROP_PUBLIC_HEALTH_EXP_LAGGED2)),10],2)`; e  menor em  `r base_efi_poor[which(base_efi_poor$PROP_PUBLIC_HEALTH_EXP_LAGGED2 == min(base_efi_poor$PROP_PUBLIC_HEALTH_EXP_LAGGED2)),1]`, `r round(base_efi_poor[which(base_efi_poor$PROP_PUBLIC_HEALTH_EXP_LAGGED2 == min(base_efi_poor$PROP_PUBLIC_HEALTH_EXP_LAGGED2)),10],2)`. (Figura 1)




```{r fig.width=12, fig.height=12}

plot1 <- ggplot(base_efi_poor)+
   geom_point(aes(PRED_NEO,reorder(LOCATION, PRED_NEO), colour = sig_neo))+
   geom_segment(aes(x = IC_05_NEO, xend = IC_95_NEO, y = LOCATION, yend = LOCATION, colour = sig_neo))+
   geom_vline(xintercept = 0, linetype = "dashed", colour = "black")+
   labs(x = "Effect neo", y = "")+
   theme_bw()+
   scale_y_discrete(position = "left")+
   scale_colour_manual(values = myPalette_rev(2), aesthetics = "colour", "Significant")
plot2 <- ggplot(base_efi_poor)+
   geom_point(aes(PRED_NEO_U5,reorder(LOCATION, PRED_NEO_U5),, color = sig_neo_u5))+
   geom_segment(aes(x = IC_05_NEO_U5, xend = IC_95_NEO_U5, y = LOCATION, yend = LOCATION, color = sig_neo_u5))+
   geom_vline(xintercept = 0, linetype = "dashed", colour = "black")+
   labs(x = "Effect 28d>5y", y = "")+
   theme_bw()+
   scale_y_discrete(position = "right")+
   scale_colour_manual(values = myPalette_rev(2), aesthetics = "colour")
ggarrange(plot1, plot2, ncol = 2, labels = c("A", "B"),legend = "bottom", common.legend = T)

```





```{r }
# Create a table of the effect of the treatment
base_efi_table <- base_efi_poor
base_efi_table$IC_NEO <- paste0("(CI95% ",round(base_efi_table$IC_05_NEO,7), " ", round(base_efi_table$IC_95_NEO,7),")")
base_efi_table$IC_NEO_U5 <- paste0("(CI95% ",round(base_efi_table$IC_05_NEO_U5,7), " ", round(base_efi_table$IC_95_NEO_U5,7),")")
base_efi_table$PRED_NEO <- round(base_efi_table$PRED_NEO,7)
base_efi_table$PRED_NEO_U5 <- round(base_efi_table$PRED_NEO_U5,7)
base_efi_table$PRED_NEO <- paste0(base_efi_table$PRED_NEO," ",base_efi_table$IC_NEO)
base_efi_table$PRED_NEO_U5 <- paste0(base_efi_table$PRED_NEO_U5," ",base_efi_table$IC_NEO_U5)
base_efi_table$MEAN_RATE_NEO <- round(base_efi_table$MEAN_RATE_NEO,3)
base_efi_table$MEAN_RATE_NEO_U5 <- round(base_efi_table$MEAN_RATE_NEO_U5,3)
base_efi_table$PUBLIC_EXP_LAGGED2 <- round(base_efi_table$PUBLIC_EXP_LAGGED2,2)
base_efi_table$PUBLIC_EXP_LAGGED2 <- paste0("US$",base_efi_table$PUBLIC_EXP_LAGGED2)
             

base_efi_table <- base_efi_table %>%
     dplyr::select("LOCATION", "PUBLIC_EXP_LAGGED2","PROP_PUBLIC_HEALTH_EXP_LAGGED2", 
                   "MEAN_RATE_NEO","PRED_NEO", "IMPACT_NEO", "EFI_NEO",
                   "MEAN_RATE_NEO_U5", "PRED_NEO_U5", "IMPACT_NEO_U5","EFI_NEO_U5")
                    
                      
                    


ggplot(subset(base_efi_table, base_efi_table$IMPACT_NEO != 0))+
   geom_col(aes(reorder(LOCATION, IMPACT_NEO),MEAN_RATE_NEO), fill = "#9E0142", position = "stack")+ 
   geom_col(aes(reorder(LOCATION, IMPACT_NEO),IMPACT_NEO), fill = "#5E4FA2", position = "stack")+
   coord_flip()+
   theme_bw()

ggplot(subset(base_efi_table, base_efi_table$IMPACT_NEO_U5 != 0))+
   geom_col(aes(reorder(LOCATION, IMPACT_NEO_U5),MEAN_RATE_NEO_U5), fill = "#9E0142", position = "stack")+ 
   geom_col(aes(reorder(LOCATION, IMPACT_NEO_U5),IMPACT_NEO_U5), fill = "#5E4FA2", position = "stack")+
   coord_flip()+
   theme_bw()

names(base_efi_table) <- c("Country", "Mean Public Exp", "Prop Public Exp to Health",
                           "Mean Neo", "Cost-Effect Neo", "Impact Neo", "Efficiency Neo",
                           "Mean 28d>5y", "Cost-Effect 28d>5y", "Impact 28d>5y", "Efficiency 28d>5y")

knitr::kable(base_efi_table, digits = 7)

```



```{r fig.width=22, fig.height=12}

effect_effici_neo <- subset(base_efi_poor, !is.na(base_efi_poor$EFI_NEO))

plot1 <- ggplot(effect_effici_neo,aes(EFI_NEO, IMPACT_NEO, colour = effect_effici_neo$INCOME_CLASS))+
            geom_text(label = effect_effici_neo$LOCATION,check_overlap = F, size =3)+
            geom_vline(aes(xintercept = quantile(effect_effici_neo$EFI_NEO,.5)))+
            geom_hline(aes(yintercept = quantile(effect_effici_neo$IMPACT_NEO,.5)))+
            ylim(c(0,15))+
            labs(y = "Impact neo", x = "Efficiency neo")+
            theme_bw()+
            scale_colour_manual(values = myPalette_rev(2), aesthetics = "colour", "Income class")

effect_effici_neo_u5 <- subset(base_efi_poor, !is.na(base_efi_poor$EFI_NEO_U5))

plot2 <- ggplot(effect_effici_neo_u5,aes(EFI_NEO_U5, IMPACT_NEO_U5, colour = effect_effici_neo_u5$INCOME_CLASS))+
            geom_text(label = effect_effici_neo_u5$LOCATION,check_overlap = F, size =3)+
            geom_vline(aes(xintercept = quantile(effect_effici_neo_u5$EFI_NEO_U5,.5)))+
            geom_hline(aes(yintercept = quantile(effect_effici_neo_u5$IMPACT_NEO_U5,.5)))+
            ylim(c(0,15))+
            labs(y = "Impact 28d>5y", x = "Efficiency 28d>5y")+
            theme_bw()+
            scale_colour_manual(values = myPalette_rev(2), aesthetics = "colour", "Income class")

ggarrange(plot1, plot2, ncol = 2, labels = c("A", "B"),legend = "bottom",common.legend = T)
```




#Proportion of Public Health Spending and Efectiveness
##Treatm Neo
```{r }

treatmneo <- subset(base_efi_poor, base_efi_poor$sig_neo == "yes")

asso_treatmneo <- lm(PRED_NEO~PROP_PUBLIC_HEALTH_EXP_LAGGED2,treatmneo) %>% summary()
beta_treatmneo <- round(asso_treatmneo$coefficients[2,1],5)
ic05_treatmneo <- round(beta_treatmneo - 1.96*asso_treatmneo$coefficients[2,2],5)
ic95_treatmneo <- round(beta_treatmneo + 1.96*asso_treatmneo$coefficients[2,2],5)

treatmneo_prop <- paste0(beta_treatmneo, " (CI95% ", ic05_treatmneo, " ", ic95_treatmneo, ")")

```


##Treatm 28d>5y
```{r }

treatmneo_u5 <- subset(base_efi_poor, base_efi_poor$sig_neo_u5 == "yes")

asso_treatmneo_u5 <- lm(PRED_NEO_U5~PROP_PUBLIC_HEALTH_EXP_LAGGED2,treatmneo_u5) %>% summary()
beta_treatmneo_u5 <- round(asso_treatmneo_u5$coefficients[2,1],5)
ic05_treatmneo_u5 <- round(beta_treatmneo_u5 - 1.96*asso_treatmneo_u5$coefficients[2,2],5)
ic95_treatmneo_u5 <- round(beta_treatmneo_u5 + 1.96*asso_treatmneo_u5$coefficients[2,2],5)

treatmneo_u5_prop <- paste0(beta_treatmneo_u5, " (CI95% ", ic05_treatmneo_u5, " ", ic95_treatmneo_u5, ")")

```

##Impact Neo
```{r }

impneo <- subset(base_efi_poor, base_efi_poor$sig_neo == "yes")

asso_impneo <- lm(IMPACT_NEO~PROP_PUBLIC_HEALTH_EXP_LAGGED2,impneo) %>% summary()
beta_impneo <- round(asso_impneo$coefficients[2,1],5)
ic05_impneo <- round(beta_impneo - 1.96*asso_impneo$coefficients[2,2],5)
ic95_impneo <- round(beta_impneo + 1.96*asso_impneo$coefficients[2,2],5)

impneo_prop <- paste0(beta_impneo, " (CI95% ", ic05_impneo, " ", ic95_impneo, ")")

```



##Impact 28d<5y
```{r }
impneo_u5 <- subset(base_efi_poor, base_efi_poor$sig_neo_u5 == "yes")

asso_impneo_u5 <- lm(IMPACT_NEO_U5~PROP_PUBLIC_HEALTH_EXP_LAGGED2,impneo_u5) %>% summary()
beta_impneo_u5 <- round(asso_impneo_u5$coefficients[2,1],5)
ic05_impneo_u5 <- round(beta_impneo_u5 - 1.96*asso_impneo_u5$coefficients[2,2],5)
ic95_impneo_u5 <- round(beta_impneo_u5 + 1.96*asso_impneo_u5$coefficients[2,2],5)

impneo_u5_prop <- paste0(beta_impneo_u5, " (CI95% ", ic05_impneo_u5, " ", ic95_impneo_u5, ")")

```




##Efficiency Neo
```{r }
efi_neo <- subset(base_efi_poor, !is.na(base_efi_poor$EFI_NEO))

asso_efi_neo <- lm(EFI_NEO ~ PROP_PUBLIC_HEALTH_EXP_LAGGED2, efi_neo) %>% summary()
beta_efi_neo <- round(asso_efi_neo$coefficients[2,1],5)
ic05_efi_neo <- round(beta_efi_neo - 1.96*asso_efi_neo$coefficients[2,2],5)
ic95_efi_neo <- round(beta_efi_neo + 1.96*asso_efi_neo$coefficients[2,2],5)

efi_neo_prop <- paste0(beta_efi_neo, " (CI95% ", ic05_efi_neo, " ", ic95_efi_neo, ")")

```

##Efficiency 28d>5y
```{r }
efi_neo_u5 <- subset(base_efi_poor, !is.na(base_efi_poor$EFI_NEO_U5))

asso_efi_neo_u5 <- lm(EFI_NEO_U5 ~ PROP_PUBLIC_HEALTH_EXP_LAGGED2, efi_neo_u5) %>% summary()
beta_efi_neo_u5 <- round(asso_efi_neo_u5$coefficients[2,1],5)
ic05_efi_neo_u5 <- round(beta_efi_neo_u5 - 1.96*asso_efi_neo_u5$coefficients[2,2],5)
ic95_efi_neo_u5 <- round(beta_efi_neo_u5 + 1.96*asso_efi_neo_u5$coefficients[2,2],5)

efi_neo_u5_prop <- paste0(beta_efi_neo_u5, " (CI95% ", ic05_efi_neo_u5, " ", ic95_efi_neo_u5, ")")

```


```{r fig.width=16, fig.height=8}
plot1 <- ggplot(treatmneo, aes(PROP_PUBLIC_HEALTH_EXP_LAGGED2,PRED_NEO))+
            geom_jitter(aes(color = LOCATION), size = 3)+
            geom_smooth(method = "lm", bs="cr", se = F, color = "#5E4FA2")+
            labs(fill="", x="Prop health exp", y="Effect neo")+
            theme_bw()+ 
            ylim(c(-.016,-.002))+
            theme(legend.position = "none") 
plot2 <- ggplot(treatmneo_u5, aes(PROP_PUBLIC_HEALTH_EXP_LAGGED2,PRED_NEO_U5))+
            geom_jitter(aes(color = LOCATION), size = 3)+
            geom_smooth(method = "lm", se = F, color = "#5E4FA2")+
            labs(fill="", x="Prop health exp", y="Effect 28d>5y")+
            theme_bw()+ 
            ylim(c(-.016,-.002))+
            theme(legend.position = "none")
plot3 <- ggplot(impneo, aes(PROP_PUBLIC_HEALTH_EXP_LAGGED2,IMPACT_NEO))+
            geom_jitter(aes(color = LOCATION), size = 3)+
            geom_smooth(method = "lm", bs="cr", se = F, color = "#5E4FA2")+
            labs(fill="", x="Prop health exp", y="Impact neo")+
            theme_bw()+ 
            ylim(c(0,15))+
            theme(legend.position = "none") 
plot4 <- ggplot(impneo_u5, aes(PROP_PUBLIC_HEALTH_EXP_LAGGED2,IMPACT_NEO_U5))+
            geom_jitter(aes(color = LOCATION), size = 3)+
            geom_smooth(method = "lm", se = F, color = "#5E4FA2")+
            labs(fill="", x="Prop health exp", y="Impact 28d>5y")+
            theme_bw()+ 
            ylim(c(0,15))+
            theme(legend.position = "none")
plot5 <- ggplot(efi_neo, aes(PROP_PUBLIC_HEALTH_EXP_LAGGED2,EFI_NEO))+
            geom_jitter(aes(color = LOCATION), size = 3)+
            geom_smooth(method = "lm", se = F, color = "#5E4FA2")+
            labs(fill="", x="Prop health exp", y="Efficiency neo")+
            theme_bw()+
            theme(legend.position = "none")
plot6 <- ggplot(efi_neo_u5, aes(PROP_PUBLIC_HEALTH_EXP_LAGGED2,EFI_NEO_U5))+
            geom_jitter(aes(color = LOCATION), size = 3)+
            geom_smooth(method = "lm", se = F, color = "#5E4FA2")+
            labs(fill="", x="Prop health exp", y="Efficiency 28d>5y")+
            theme_bw()+
            theme(legend.position = "none")



ggarrange(plot1, plot2,plot4, plot3, plot5, plot6, nrow = 3,ncol = 2,
          labels = c("A", "B", "C", "D", "E", "F"))

```



Associação entre efeito na mortalidade neonatal e proporção de gastos públicos destinados à saúde: `r treatmneo_prop`
Associação entre efeito na mortalidade de crianças entre 28 dias e cinco anos e proporção de gastos públicos destinados à saúde: `r treatmneo_u5_prop`

Associação entre impacto na mortalidade neonatal e proporção de gastos públicos destinados à saúde: `r impneo_prop`
Associação entre impacto na mortalidade de crianças entre 28 dias e cinco anos e proporção de gastos públicos destinados à saúde: `r impneo_u5_prop`
Associação entre eficiência na redução da mortalidade neonatal e proporção de gastos públicos destinados à saúde: `r efi_neo_prop`
Associação entre eficiência na redução da mortalidade de crianças entre 28 dias e cinco anos e proporção de gastos públicos destinados à saúde: `r efi_neo_u5_prop`
